+++ src/transport/grpc/store.rs
@@ -36,7 +36,7 @@ use crate::{
         OptStoredProp, PropertyInterface, PropertyMapping, PropertyStore, StoreCapabilities,
         StoredProp,
     },
-    AstarteType,
+    AstarteData,
 };
 
 use super::{
@@ -108,7 +108,7 @@ impl StoreCapabilities for GrpcStore {
 impl PropertyStore for GrpcStore {
     type Err = GrpcStoreError;
 
-    async fn store_prop(&self, _prop: StoredProp<&str, &AstarteType>) -> Result<(), Self::Err> {
+    async fn store_prop(&self, _prop: StoredProp<&str, &AstarteData>) -> Result<(), Self::Err> {
         // do not store properties locally when connected as a message hub node
         Ok(())
     }
@@ -117,7 +117,7 @@ impl PropertyStore for GrpcStore {
         &self,
         property: &PropertyMapping<'_>,
         _interface_major: i32,
-    ) -> Result<Option<AstarteType>, Self::Err> {
+    ) -> Result<Option<AstarteData>, Self::Err> {
         let property = self
             .client
             .lock()
@@ -132,7 +132,7 @@ impl PropertyStore for GrpcStore {
 
         property
             .data
-            .map(|data| AstarteType::try_from(data).map_err(GrpcStoreError::from))
+            .map(|data| AstarteData::try_from(data).map_err(GrpcStoreError::from))
             .transpose()
     }
 
@@ -207,7 +207,7 @@ mod test {
     use crate::interface::Ownership;
     use crate::store::PropertyMapping;
     use crate::store::StoredProp;
-    use crate::AstarteType;
+    use crate::AstarteData;
     use crate::Interface;
 
     #[tokio::test]
@@ -343,7 +343,7 @@ mod test {
 
     #[tokio::test]
     async fn test_grpc_store_device_prop_not_stored() {
-        let inner_value = AstarteType::Integer(1);
+        let inner_value = AstarteData::Integer(1);
         const PATH: &str = "/path1";
         let server_interface = "com.server.interface";
         let server_prop = StoredProp {